Separate Dapr constants from IT container constants
