+++ testcontainers-dapr/src/main/java/io/dapr/testcontainers/DaprContainerConstants.java
@@ -2,7 +2,7 @@ package io.dapr.testcontainers;
 
 public interface DaprContainerConstants {
   String DAPR_VERSION = "1.15.4";
+  String DAPR_RUNTIME_IMAGE_TAG = "daprio/daprd:" + DAPR_VERSION;
   String DAPR_PLACEMENT_IMAGE_TAG = "daprio/placement:" + DAPR_VERSION;
   String DAPR_SCHEDULER_IMAGE_TAG = "daprio/scheduler:" + DAPR_VERSION;
-  String DAPR_RUNTIME_IMAGE_TAG = "daprio/daprd:" + DAPR_VERSION;
 }
